Car Breakdown
Yuki has received a request to pick up a car which has broken down nearby. Since Yuki's company has a tie up with the car rental company to service the cars periodically or in an event of a breakdown. 
Nick must arrange for a replacement car as the customer's car has broken down during the rental period. He also needs to alert the car service company to tow the car and get it fixed 
Anja must note down complaints from the customers regarding the cleanliness of the car, issues with it etc when the rental period starts