Car booking
Max wants to book a rental car for a specific period as his own car has gone to service for some overhaul.
Max wants the flexibility to change or alter the dates of his booking. Since he needs to go to the company headquarters once a quarter for meetings due to the nature of his work and the dates are not finalized in advance.  
Max wants the ability to quickly cancel his booking of the rental car. This scenario can happen if he must take a few days off from work for personal reasons. 
Max wants the rental company to provide him with a different rental car in case the one he had booked has broken down. Since he must be on the road the whole day long and he would not want to cancel his meetings due to the breakdown.